Moving all medical facilities (hospitals, drive-thru testing site) to be handled by the same class since they share a lot of functionality.
Also merged the hospital and map_event data into a single file.
Leaving map-events class as is in case we want to use it to display other things.
Added a file that handles reading parameter files. This is an attempt to make the project more generic.
The color and size of the medical facility circles is now in a csv file.
Added icons for medical facility type.
Added a .gitignore to the visualization-ui submodule